Add basic life cycle events
This adds Event::LifeCycle, and the LifeCycle enum, which for now
has two variants: ApplicationStarted and WindowConnected.

The first of these is sent only once, when the application starts;
the second is sent to a window each time a window is added.

ApplicationLaunched is kind of fiddly; if we want to send that message
to a window, we need to track which windows exist at launch time,
and then send it to each of those windows when they actually connect.

It may be that we only really want to send this message to the top-level
application delegate, which would require tracking less state and
be generally cleaner.